consolidates skin tweaks into one class and fixes green number not working
