+++ src/hooks/hooks.cpp
@@ -2,7 +2,6 @@
 #include "hooks/srcnumber.h"
 #include "hooks/mirror.h"
 #include "hooks/cursor.h"
-#include "hooks/drawnum.h"
 #include "hooks/updategamestate.h"
 #include "hooks/judgement.h"
 #include "hooks/loadbms.h"
@@ -12,7 +11,6 @@ void hooks::Setup()
 {
     cursor::Install();
     updategamestate::Install();
-    drawnum::Install();
     loadbms::Install();
     input::Install();
 }
