+++ src/hooks/hooks.h
@@ -3,7 +3,7 @@
 #include "hooks/judgement.h"
 #include "hooks/srcnumber.h"
 #include "hooks/mirror.h"
-#include "hooks/skin.h"
+#include "hooks/skinmisc.h"
 
 namespace hooks {
     void Setup();
@@ -12,6 +12,7 @@ namespace hooks {
     inline ReplayFix replay_fix;
     inline SrcNumber src_number;
     inline JudgementProcessing judgement_processing;
+    inline GreenNumber green_number;
     inline Mirror mirror;
     inline SkinMisc skin_misc;
 }
